+++ microdata/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.3.0.dev'
+__version__ = '0.3.0'
 __description__ = 'Microdata semantic markups support for Pelican Blog Generator'